+++ src/item.rs
@@ -1,0 +1,16 @@
+#[allow(dead_code)]
+pub enum Item {
+    First,
+    Second,
+    Third,
+}
+
+impl Item {
+    pub fn index(&self) -> usize {
+        match self {
+            Item::First => 0,
+            Item::Second => 1,
+            Item::Third => 2,
+        }
+    }
+}
